backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
Pillow
requests
aiofiles
